+++ changelog.md
@@ -4,6 +4,7 @@
 
 - structurizr-dsl: Allows archetypes to be used via workspace extension.
 - structurizr-dsl: Fixes https://github.com/structurizr/java/issues/392 (SVG not supported in base 64 encoding not mentioned in documentation).
+- structurizr-export: Adds support for rank and node separation to the StructurizrPlantUMLExporter.
 
 ## v4.0.0 (28th March 2025)